+++ examples/generateExamples.ts
@@ -224,7 +224,15 @@ const generateExampleFromTemplate = async (exampleID: string) => {
 };
 
 // #region Other examples' package.json
-const otherExamples = ['base-ui', 'ci', 'native', 'next', 'preact', 'tremor'] as const;
+const otherExamples = [
+  'base-ui',
+  'ci',
+  'native',
+  'next',
+  'preact',
+  'rules-engine',
+  'tremor',
+] as const;
 
 const updateOtherExample = async (otherExampleName: string) => {
   const otherExamplePkgJSON: PackageJSON = await Bun.file(