+++ utils/devapp/useDevApp.ts
@@ -19,7 +19,7 @@ import {
 } from './constants';
 import type { CommonRQBProps, DemoOptions } from './types';
 import type { OptionsAction } from './utils';
-import { generatePermalinkHash, getFormatQueryString, optionsReducer } from './utils';
+import { generatePermalinkHash, getFormatQueryString, generateOptionsReducer } from './utils';
 
 Object.defineProperty(globalThis, 'RQButils', { value: RQButils });
 
@@ -31,8 +31,11 @@ const getOptionsFromHash = () =>
 // Initialize options from URL hash
 const initialOptionsFromHash = getOptionsFromHash();
 
+const emptyObject = {};
+
 export const useDevApp = <ExtraOptions extends Record<string, boolean | undefined>>(
-  extraOptions: Partial<ExtraOptions> = {}
+  extraOptions: ExtraOptions = emptyObject as ExtraOptions,
+  exportFormats: Record<string, string> = emptyObject
 ): {
   actions: [string, () => void][];
   commonRQBProps: CommonRQBProps;
@@ -45,9 +48,11 @@ export const useDevApp = <ExtraOptions extends Record<string, boolean | undefine
   updateOptions: React.ActionDispatch<
     [action: OptionsAction<Partial<ExtraOptions> & { [k: string]: boolean }>]
   >;
+  optionsReducer: ReturnType<typeof generateOptionsReducer<ExtraOptions>>;
 } => {
   const [query, setQuery] = useState(initialQuery);
   const [queryIC, setQueryIC] = useState(initialQueryIC);
+  const optionsReducer = useMemo(() => generateOptionsReducer(extraOptions), [extraOptions]);
   const [optVals, updateOptions] = useReducer(optionsReducer, {
     ...defaultOptions,
     ...extraOptions,
@@ -71,16 +76,25 @@ export const useDevApp = <ExtraOptions extends Record<string, boolean | undefine
 
   const formatQueryResults = useMemo(
     () =>
-      formatMap.map(([format]) => {
-        const formatQueryOptions: FormatQueryOptions = {
-          format,
-          fields: optVals.validateQuery ? fields : undefined,
-          parseNumbers: optVals.parseNumbers,
-        };
-        const q = optVals.independentCombinators ? queryIC : query;
-        return [format, getFormatQueryString(q, formatQueryOptions)] as const;
-      }),
-    [optVals.validateQuery, optVals.parseNumbers, optVals.independentCombinators, queryIC, query]
+      Object.keys(exportFormats).length > 0
+        ? (Object.entries(exportFormats) as (readonly [ExportFormat, string])[])
+        : formatMap.map(([format]) => {
+            const formatQueryOptions: FormatQueryOptions = {
+              format,
+              fields: optVals.validateQuery ? fields : undefined,
+              parseNumbers: optVals.parseNumbers,
+            };
+            const q = optVals.independentCombinators ? queryIC : query;
+            return [format, getFormatQueryString(q, formatQueryOptions)] as const;
+          }),
+    [
+      exportFormats,
+      optVals.validateQuery,
+      optVals.parseNumbers,
+      optVals.independentCombinators,
+      queryIC,
+      query,
+    ]
   );
 
   const actions = useMemo(
@@ -123,7 +137,7 @@ export const useDevApp = <ExtraOptions extends Record<string, boolean | undefine
           },
         ],
       ] satisfies [string, () => void][],
-    [optVals]
+    [optVals, optionsReducer]
   );
 
   const onQueryChange = useCallback((q: RuleGroupType) => setQuery(q), []);
@@ -138,6 +152,7 @@ export const useDevApp = <ExtraOptions extends Record<string, boolean | undefine
     optVals,
     query,
     queryIC,
+    optionsReducer,
     updateOptions,
   };
 };